+++ test/install_generator_test.rb
@@ -22,6 +22,7 @@ class InstallGeneratorTest < Minitest::Test
     assert_includes initializer, "Route configuration must load before Rails draws routes."
     assert_includes initializer, "additional_profile_attributes"
     assert_includes initializer, "require_password_confirmation"
+    assert_includes initializer, "login_title"
     assert_includes initializer, "omniauth_providers"
     assert_includes initializer, "omniauth_create_account"
     assert_includes initializer, "recording_studio_user/omniauth_callbacks"